cambios menores
Se corrigen algunos errores en cuanto a las acciones de algunos roles, campos de texto, entre otros.

+++ resources/views/supervisores/table.blade.php
@@ -11,7 +11,9 @@
 
                         <div class="box-tools pull-right">
                             @if(Auth::user()->hasRole(['Super Admin', 'Supervisor 1']))
-                                {!! Form::open(['route' => ['supervisores.destroy', $supervisor->id], 'method' => 'delete']) !!}
+                                @if(!$supervisor->hasRole(['Super Admin', 'Supervisor 1']))
+                                    {!! Form::open(['route' => ['supervisores.destroy', $supervisor->id], 'method' => 'delete']) !!}
+                                @endif
                             @endif    
                                 <div class='btn-group'>
 
@@ -25,16 +27,22 @@
                                     @endcomponent
 
                                     <!-- Boton para editar los datos del encuestador -->
-                                    {{-- @if (Auth::user()->hasRole('Administrador')) --}}
-                                        <a href="{!! route('supervisores.edit', [$supervisor->id]) !!}" class='btn btn-default btn-xs'>
-                                            <i class="glyphicon glyphicon-edit"></i>
-                                        </a>
-                    
+                                        {{-- @if(Auth::user()->id == $supervisor->id) --}}
                                         @if(Auth::user()->hasRole(['Super Admin', 'Supervisor 1']))
-                                            <!-- Boton para eliminar los datos del encuestador -->
-                                            {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs','onclick' => "return confirm('¿Está seguro de querer eliminar?')", 'data-toggle' => 'modal', 'data-target' => '#modal-danger']) !!}
+                                            @if(!$supervisor->hasRole(['Super Admin', 'Supervisor 1']))
+                                                <a href="{!! route('supervisores.edit', [$supervisor->id]) !!}" class='btn btn-default btn-xs'>
+                                                    <i class="glyphicon glyphicon-edit"></i>
+                                                </a>
+                                            @endif
+                                        {{-- @endif --}}
                                         @endif
-                                    {{-- @endif --}}
+
+                                        @if(Auth::user()->hasRole(['Super Admin', 'Supervisor 1']))
+                                            @if(!$supervisor->hasRole(['Super Admin', 'Supervisor 1']))
+                                                <!-- Boton para eliminar los datos del encuestador -->
+                                                {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs','onclick' => "return confirm('¿Está seguro de querer eliminar?')", 'data-toggle' => 'modal', 'data-target' => '#modal-danger']) !!}
+                                            @endif
+                                        @endif
 
                                     <!-- Boton para minimizar/maximiar cada cuadro -->
                                     <button type="button" class="btn btn-info btn-xs" data-widget="collapse">
@@ -43,7 +51,9 @@
         
                                 </div>
                             @if(Auth::user()->hasRole(['Super Admin', 'Supervisor 1']))
-                                {!! Form::close() !!}
+                                @if(!$supervisor->hasRole(['Super Admin', 'Supervisor 1']))
+                                    {!! Form::close() !!}
+                                @endif
                             @endif
                         </div>
                     </div>
